coloring

+++ tensorflow/core/graph/rocm/attr_to_rtg.h
@@ -25,7 +25,7 @@ namespace tensorflow {
 namespace rtglib {
 namespace convert {
 void GetProgram(const NameAttrList&, void **, int&, string);
-void EvalProgram(void*, Tensor*, std::vector<const Tensor*>&, bool, void*, int, string);
+void EvalProgram(void*, Tensor*, std::vector<const Tensor*>&, bool, void*, int64, string);
 void GetOutputShape(void *, TensorShape&);
 void AdjustShape( void *, std::vector<const Tensor*>&, string);
